requirements.txt
openwa==0.4.0
pyttsx3==0.1.1
pillow==10.4.0
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6